The ring task's solution from Joe Armstrong's Programming Erlang book
There are two variations in this commit:
- one that runs processes on the same node (computer)
- one that runs processes on the current node and another one
Each version creates a ring of N processes and sends a message through the
ring M messages.
Then the ring is stopped by sending through the ring a stop signal.

The task is on p157 in the first edition (available at pragprog.com).